Add files via upload
In the Data of Madhav Store we had extract the data according to get the insides from the data about the Profit and sales of the store according to the Category of Items, Sub Category of the Items. The main occupation is to get the answers of what sales the most at which region, At what month of it and of how many in Quantity through which Customer.